fixes timeline chart
