+++ src/components/Proposals/ProposalPage/OPProposalPage/ProposalVotesSummaryDetails/ProposalVotesSummaryDetails.tsx
@@ -1,5 +1,5 @@
 import Image from "next/image";
-import { HStack, VStack } from "@/components/Layout/Stack";
+import { VStack } from "@/components/Layout/Stack";
 
 import infoTransparentIcon from "@/icons/info-transparent.svg";
 import checkIcon from "@/icons/check.svg";
@@ -10,28 +10,10 @@ import TokenAmountDisplay from "@/components/shared/TokenAmountDisplay";
 import { ParsedProposalResults } from "@/lib/proposalUtils";
 import { format } from "date-fns";
 import Link from "next/link";
-import { formatNumber } from "@/lib/utils";
+import { formatNumber, formatNumberWithScientificNotation } from "@/lib/utils";
 
 import Tenant from "@/lib/tenant/tenant";
 const { token } = Tenant.current();
-
-function formatNumberWithScientificNotation(x: any) {
-  if (Math.abs(x) < 1.0) {
-    var e = parseInt(x.toString().split("e-")[1]);
-    if (e) {
-      x *= Math.pow(10, e - 1);
-      x = "0." + new Array(e).join("0") + x.toString().substring(2);
-    }
-  } else {
-    var e = parseInt(x.toString().split("+")[1]);
-    if (e > 20) {
-      e -= 20;
-      x /= Math.pow(10, e);
-      x += new Array(e + 1).join("0");
-    }
-  }
-  return x;
-}
 
 export default function ProposalVotesSummaryDetails({
   proposal,
